✨ Split splash master to separate in rive, video and lottie
