+++ example/android/app/src/main/res/values-night/styles.xml
@@ -4,7 +4,7 @@
   <style name="LaunchTheme" parent="@android:style/Theme.Black.NoTitleBar">
     <!-- Show a splash screen on the activity. Automatically removed when
              the Flutter engine draws its first frame -->
-    <item name="android:windowBackground">@drawable/splash_screen_dark</item>
+    <item name="android:windowBackground">@drawable/splash_screen</item>
   </style>
   <!-- Theme applied to the Android Window as soon as the process has started.
          This theme determines the color of the Android Window while your
@@ -13,6 +13,6 @@
 
          This Theme is only used starting with V2 of Flutter's Android embedding. -->
   <style name="NormalTheme" parent="@android:style/Theme.Black.NoTitleBar">
-    <item name="android:windowBackground">@drawable/splash_screen_dark</item>
+    <item name="android:windowBackground">@drawable/splash_screen</item>
   </style>
 </resources>
