+++ lib/splash_master.dart
@@ -21,8 +21,5 @@
  */
 
 export 'package:splash_master/core/source.dart';
-export 'package:splash_master/core/splash_master.dart';
 export 'package:splash_master/enums/splash_master_enums.dart';
-export 'package:splash_master/splashes/lottie/lottie_config.dart';
-export 'package:splash_master/splashes/video/video_config.dart';
-export 'package:splash_master/splashes/rive/rive_config.dart';
+export 'package:splash_master/core/utils.dart';
