+++ packages/splash_master_rive/example/pubspec.yaml
@@ -1,0 +1,18 @@
+name: splash_master_rive_example
+description: "Example app demonstrating the splash_master_rive package."
+publish_to: none
+
+environment:
+  sdk: '>=3.4.3 <4.0.0'
+
+dependencies:
+  flutter:
+    sdk: flutter
+  splash_master_rive:
+    path: ../
+
+flutter:
+  uses-material-design: true
+
+  assets:
+    - assets/
