+++ packages/splash_master_rive/pubspec.yaml
@@ -1,0 +1,18 @@
+name: splash_master_rive
+description: "Rive animation splash screen widget for splash_master."
+version: 2.0.0
+publish_to: none
+
+environment:
+  sdk: '>=3.4.3 <4.0.0'
+  flutter: '>=3.3.0'
+
+dependencies:
+  flutter:
+    sdk: flutter
+  splash_master:
+    path: ../../
+  rive: ^0.13.20
+
+dev_dependencies:
+  flutter_lints: ^3.0.0
