+++ packages/splash_master_video/lib/splash_master_video.dart
@@ -1,0 +1,23 @@
+/// Video splash screen widget for splash_master.
+///
+/// Add to your `pubspec.yaml`:
+/// ```yaml
+/// dependencies:
+///   splash_master_video:
+///     path: ../packages/splash_master_video  # adjust path as needed
+/// ```
+///
+/// Then use [VideoSplash] in your Flutter app:
+/// ```dart
+/// import 'package:splash_master_video/splash_master_video.dart';
+///
+/// VideoSplash(
+///   source: AssetSource('assets/splash.mp4'),
+///   nextScreen: const MyApp(),
+/// )
+/// ```
+library splash_master_video;
+
+export 'package:splash_master/splash_master.dart';
+export 'src/video_config.dart';
+export 'src/video_splash.dart';
