+++ pubspec.yaml
@@ -17,9 +17,6 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  lottie: ^3.1.2
-  video_player: ^2.9.1
-  rive: ^0.13.20
   args: ^2.3.0
   xml: ^6.5.0
   yaml: ^3.1.2
